Move types from `core` to separate file
